Fix TP_WrapText crashing when scrolling
Yes, we definitely needed to "do something" here! In order to scroll the text up, you have to grab the block of screen pixels below the first line and copy then copy them over it. That's what VL_ScreenToScreen was doing, but since we have VW_ScreenToMem we can just copy the screen pixels to a buffer and then redraw them with VW_MemToScreen. There might be a simpler solution staring me in the face, but at least this seems to work fine.

However, there are some problems with integer overflow that forced me to change the size & signedness of certain variables. This needs revisiting and fixing properly, because it could cause problems later on (and it's a terrible practice anyway).

+++ id_vw.c
@@ -10,7 +10,7 @@ pictabletype *pictable;
 
 uint32_t     *ylookup;
 
-int          px,py;
+int16_t      px,py;
 int          fontcolor,backcolor;
 int          fontnumber;
